conf.py: remove latex font modules
The latex font modules don't seem to work very well on linux build
systems, including the GH builder. The code blocks end up being a serif
font, rather than the monospace fonts they are supposed to be. The
modules don't actually seem to be necessary, as all the fonts default to
sane values.

Remove the font modules and allow the latex defaults to be used.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>