appendix_2.rst: denote variables and commands as code
While these variables are only used notionally (e.g. not even in pseudocode),
formatting them in book script looks distinctly out of place.

Note that the sample keys are not similarly formatted because of their
truncation. If the ellipses were inside the code block, it might give the false
impressiont that WireGuard keys literally do contain ellipses; but if they were
outside the code block, there would be a confusing amount of excess space
between the truncated key and the truncated keys and the ellipses. Leaving them
formatted in book script is not as jarring as for the pseudo-variables and
avoids the problem.

This change also formats several commands in the same table in monospace.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>