appendix_3.rst: style fixes
Named firewalld entities, network interfaces, package names, and paths should be
in monospace.

Argument snippets should be in monospace even if they can be understood
transparently: "UDP port 2222" should be in book text, but `2222/udp` is
something that literally belongs on the command line.

Wrap code snippets so that sphinx doesn't try to wrap them for us.

"INGRESS" and "EGRESS" are not named firewalld entities and should be
capitalized normally.

Code blocks should be formatted appropriately.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>