conf.py: reduce pointsize from 12pt to 11pt
12pt is significantly larger than comparable documents, notably NIST standards
and past NI manuals in PDF format. It is alone responsible for quite a few
typesetting bugs due to running out of space. Reducing it to 11pt significantly
improves typesetting quality at negligible or perhaps even positive change to
legibility.

That said, this caused the already-scaled-down Appendix 2 table to be too small
to read, so its \scalefont has been adjusted appropriately.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>