.github: build project on pr checks
The current pr checks only test that commits are signed-off. A different
workflow builds the PDF on pushes into master, but also floats its
python dependencies to the latest version in the host system - which
can cause discrepancies in the PDF output between developer builds and
the pipeline.

* Remove the sign-off check from the PR, since it is more trouble than it
is worth.
* Remove the master push workflow and instead build the project on all
  PR checks.
* Modify the build action to install python requirements from the
  requirements.txt file.

Fixes: https://github.com/ni/nilrt-snac-guide/issues/13
Signed-off-by: Alex Stewart <alex.stewart@emerson.com>