document_metadata: remove table around contact info
The table element surrounding the document contact info isn't really
needed for clarity and states some obvious context information.

Convert the contact info to a simple block quote for clarity and
concision. Also move the contact email into this section to free up the
bottom of the section for the changelog.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>